# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: rosshield API
  version: 0.0.1
  summary: ROS2 Fleet Security Audit Platform API
  description: |
    Skeleton spec for Phase 0. Contract-first — server handlers, client SDKs,
    and CLI command shapes are generated from this document. See
    `docs/design/05-api-and-auth.md` for the full design.

    Status: v1 under design. No operation here is considered stable.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://127.0.0.1:8080
    description: Local desktop / development
  - url: https://{host}
    description: On-prem or appliance
    variables:
      host:
        default: rosshield.example.internal
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/v1/robots:
    get:
      tags: [robot]
      operationId: listRobots
      summary: List robots
      x-status: todo
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Sort"
        - { in: query, name: fleetId, schema: { type: string } }
        - { in: query, name: tag,     schema: { type: string } }
        - { in: query, name: q,       schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListRobotsResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [robot]
      operationId: createRobot
      summary: Register a robot
      description: |
        Accepts plaintext credentials (memory-only) — domain layer wraps via KEK→DEK.
        Response excludes plaintext credentials (Robot meta + credentialId only).
        Conflicts (409): duplicate name within fleet, duplicate host:port.
      x-status: todo
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRobotRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRobotResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /api/v1/reports:
    get:
      tags: [reporting]
      operationId: listReports
      summary: List reports
      description: |
        Lists generated reports for the current tenant. PDF body is not included —
        only metadata + signature presence flag. Filter by `sessionId` to scope to a
        single scan.
      x-status: phase1
      parameters:
        - { in: query, name: sessionId, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListReportsResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by `/api/v1/auth/login` or OIDC/SAML callback
    apiKeyAuth:
      type: http
      scheme: bearer
      description: "API key in `Authorization: Bearer fg_live_...` form"
  parameters:
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
    Sort:
      in: query
      name: sort
      description: "Field name; prefix with `-` for descending."
      schema: { type: string }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      description: ULID or UUID. Same key within 24h returns the original response.
      schema: { type: string, minLength: 16, maxLength: 64 }
  responses:
    ErrorResponse:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    Meta:
      type: object
      required: [requestId]
      properties:
        requestId:     { type: string }
        serverVersion: { type: string }
        responseAt:    { type: string, format: date-time }
...
    ErrorCategory:
      type: string
      enum:
        - unauthenticated
        - forbidden
        - not_found
        - conflict
        - validation
        - rate_limited
        - upstream
        - internal
    Error:
      type: object
      required: [code, category, message]
      properties:
        code:     { type: string, description: "Dotted code (e.g. `robot.not_found`)" }
        category: { $ref: "#/components/schemas/ErrorCategory" }
        message:  { type: string }
        target:
          type: object
          properties:
            type: { type: string }
            id:   { type: string }
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok:    { type: boolean, const: false }
        error: { $ref: "#/components/schemas/Error" }
        meta:  { $ref: "#/components/schemas/Meta" }

    # ─── Robot ────────────────────────────────────────────────────────────
    Robot:
      type: object
      required: [id, tenantId, fleetId, name, host, port, authType, criticality]
      properties:
        id:          { type: string }
        tenantId:    { type: string }
        fleetId:     { type: string }
        name:        { type: string }
        host:        { type: string }
        port:        { type: integer }
        authType:    { type: string, enum: [password, privateKey] }
        osDistro:    { type: string }
        rosDistro:   { type: string }
        tags:
          type: array
          items: { type: string }
        role:        { type: string }
        criticality: { type: string, enum: [low, medium, high, critical] }
    ListRobotsResponse:
      type: object
      required: [robots]
      properties:
        robots:
          type: array
          items: { $ref: "#/components/schemas/Robot" }
    CreateRobotRequest:
      type: object
      required: [fleetId, name, host, authType, username]
      properties:
        fleetId:              { type: string }
        name:                 { type: string }
        host:                 { type: string }
        port:                 { type: integer, description: "0 → domain default (22)" }
        authType:             { type: string, enum: [password, privateKey] }
        username:             { type: string }
        password:             { type: string, format: password }
        privateKeyPem:        { type: string }
        privateKeyPassphrase: { type: string, format: password }
        osDistro:             { type: string }
        rosDistro:            { type: string }
        tags:
          type: array
          items: { type: string }
        role:                 { type: string }
        criticality:          { type: string, enum: [low, medium, high, critical], description: "Empty → domain default (medium)" }
    CreateRobotResponse:
      type: object
      required: [robot, credentialId]
      properties:
        robot:        { $ref: "#/components/schemas/Robot" }
        credentialId: { type: string }

    # ─── Scan ─────────────────────────────────────────────────────────────
...
    Report:
      type: object
      required: [id, tenantId, sessionId, format, pdfSha256, pdfSizeBytes, generatedAt, generatedBy, signed]
      properties:
        id:           { type: string }
        tenantId:     { type: string }
        sessionId:    { type: string }
        format:       { type: string }
        pdfSha256:    { type: string }
        pdfSizeBytes: { type: integer, format: int64 }
        generatedAt:  { type: string, format: date-time }
        generatedBy:  { type: string }
        signed:       { type: boolean }
    ListReportsResponse:
      type: object
      required: [reports]
      properties:
        reports:
          type: array
          items: { $ref: "#/components/schemas/Report" }

    # ─── Insight ──────────────────────────────────────────────────────────
...